fix rare memset bug

+++ kernel/include/lib/string.h
@@ -4,7 +4,7 @@
 #include <lib/barelib.h>
 
 int16_t strcmp(const char*, const char*);
-void* memset(void*, byte, int32_t);
+void* memset(void*, byte, uint64_t);
 void* memcpy(void*, const void*, uint64_t);
 int16_t memcmp(const void*, const void*, uint64_t);
